crawler/requirements.txt
playwright>=1.40.0
lxml>=5.0.0
pyyaml>=6.0.0

